refactor: Filter 컴포넌트 CategorySidebar 네이밍 변경
