+++ src/pages/Explore/ui/CategorySidebar/CategorySidebar.styled.ts
